Fixing warnings

+++ JiveOne/JCKeyboardViewController.m
@@ -15,7 +15,7 @@
     if ([sender isKindOfClass:[UIButton class]])
     {
         UIButton *button = (UIButton *)sender;
-        NSString *typedChar = [self characterFromNumPadTag:button.tag];
+        NSString *typedChar = [self characterFromNumPadTag:(int)button.tag];
         
         self.outputLabel.text =  [NSString stringWithFormat:@"%@%@", self.outputLabel.text, typedChar];
